requirements.txt
cohere==5.18.0
fastapi==0.117.1
Pillow==11.3.0
Requests==2.32.5
ultralytics==8.3.159
